Fused utils/config and utils/plot into utils/json, improved documentation by adding a page for JSON configs
