+++ grapher/include/grapher/plotters/debug.hpp
@@ -4,8 +4,16 @@
 
 namespace grapher::plotters {
 
-/// \ingroup plotters
-/// Debug plotter, outputs statistics on benchmark categories
+/// Debug plotter. Outputs various statistics on benchmark categories to debug
+/// category building or traversal issues.
+///
+/// Example config:
+/// \code{.json}
+/// {
+///   "plotter": "debug"
+/// }
+/// \endcode
+
 struct plotter_debug_t : public plotter_base_t {
   void plot(benchmark_set_t const &bset, std::filesystem::path const &dest,
             grapher::json_t const &config) const override;
