+++ grapher/include/grapher/plotters/grouped_histogram.hpp
@@ -4,8 +4,6 @@
 
 namespace grapher::plotters {
 
-/// \ingroup plotters
-/// Display bars
 struct plotter_grouped_histogram_t : public plotter_base_t {
   void plot(benchmark_set_t const &bset, std::filesystem::path const &dest,
             grapher::json_t const &config) const override;
