+++ grapher/include/grapher/utils/json.hpp
@@ -9,8 +9,10 @@
 
 #include <fmt/core.h>
 
+#include <sciplot/sciplot.hpp>
+
 #include "grapher/core.hpp"
-#include "grapher/utils/config.hpp"
+#include "grapher/predicates.hpp"
 #include "grapher/utils/error.hpp"
 
 namespace grapher {
@@ -110,5 +112,50 @@ json_at_ref(grapher::json_t const &object, LocType const &field_location,
 
   return object[field_location].template get_ref<ReferenceType>();
 }
+
+// =============================================================================
+// Group descriptors
+
+/// Named set of constraint.
+struct group_descriptor_t {
+  std::string name;
+  grapher::json_t::array_t predicates;
+};
+
+/// Returns a default group descriptor as an example.
+group_descriptor_t get_default_group_descriptor();
+
+/// Extracts predicates from a group descriptor.
+std::vector<predicate_t> get_predicates(group_descriptor_t const &descriptor);
+
+grapher::json_t::array_t extract_group(group_descriptor_t const &descriptor,
+                                       grapher::json_t::array_t const &events);
+
+/// Exports a single group descriptor into a JSON object.
+grapher::json_t group_descriptor_json(group_descriptor_t const &descriptor);
+
+/// Exports a vector of group descriptors into a JSON object.
+grapher::json_t::array_t
+write_descriptors(std::vector<group_descriptor_t> const &descriptors);
+
+/// Reads a single descriptor.
+group_descriptor_t read_descriptor(grapher::json_t const &j);
+
+/// Reads descriptors from a predicate list.
+std::vector<group_descriptor_t>
+read_descriptors(grapher::json_t::array_t const &list);
+
+// =============================================================================
+// Plotter configuration
+
+/// Plot saving helper function.
+void save_plot(sciplot::Plot const &plot, std::string const &dest,
+               grapher::json_t const &config);
+
+/// Returns the default configuration for apply_config.
+grapher::json_t base_default_config();
+
+/// Apply config to plot.
+sciplot::Plot &apply_config(sciplot::Plot &plot, grapher::json_t config);
 
 } // namespace grapher
