+++ grapher/lib/grapher/plotters/stack.cpp
@@ -11,10 +11,8 @@
 #include "grapher/core.hpp"
 #include "grapher/plotters/stack.hpp"
 #include "grapher/predicates.hpp"
-#include "grapher/utils/config.hpp"
 #include "grapher/utils/error.hpp"
 #include "grapher/utils/json.hpp"
-#include "grapher/utils/plot.hpp"
 
 namespace grapher::plotters {
 
@@ -22,33 +20,20 @@ grapher::json_t plotter_stack_t::get_default_config() const {
   grapher::json_t res = grapher::base_default_config();
 
   res["plotter"] = "stack";
-
-  // Basic values, probably no need to change them
   res["value_json_pointer"] = "/dur";
-  res["name_json_pointer"] = "/name";
-
-  // Some matchers as an example...
   res["group_descriptors"] =
       write_descriptors({get_default_group_descriptor()});
 
   return res;
 }
 
-/// For each benchmark in the category, generates a stakcked curve graph where
-/// each curve corresponds to a matcher in the matchers JSON field.
-
 void plotter_stack_t::plot(benchmark_set_t const &bset,
                            std::filesystem::path const &dest,
                            grapher::json_t const &config) const {
-  // Config
+  // Config reading
 
   grapher::json_t::json_pointer feature_value_jptr(
       config.value("value_json_pointer", "/dur"));
-
-  grapher::json_t::json_pointer feature_name_jptr(
-      config.value("name_json_pointer", "/name"));
-
-  std::string plot_file_extension = config.value("plot_file_extension", ".svg");
 
   std::vector<group_descriptor_t> descriptors = read_descriptors(
       json_at_ref<json_t::array_t const &>(config, "group_descriptors"));
@@ -57,7 +42,7 @@ void plotter_stack_t::plot(benchmark_set_t const &bset,
 
   std::vector<sciplot::Plot> plots;
 
-  // Saving max y value for normalization
+  // Storing max y value for normalization
   double max_y_val = 0.;
 
   /// Draws a stacked curve graph for a given benchmark
