+++ grapher/lib/grapher/utils/json.cpp
@@ -1,7 +1,6 @@
 #include "grapher/utils/json.hpp"
 #include "grapher/core.hpp"
 #include "grapher/predicates.hpp"
-#include "grapher/utils/config.hpp"
 
 #include <algorithm>
 #include <execution>
@@ -10,6 +9,91 @@
 #include <nlohmann/json.hpp>
 
 namespace grapher {
+
+/// Default group descriptor values:
+///
+/// \code{.js}
+/// {
+///   "group_descriptors": [
+///     {
+///       "name": "All",
+///       "predicates": [
+///         {
+///           "pointer": "/name",
+///           "regex": "*",
+///           "type": "regex"
+///         }
+///       ]
+///     }
+///   ]
+/// }
+/// \endcode
+
+group_descriptor_t get_default_group_descriptor() {
+  return {.name = "All",
+          .predicates = grapher::json_t::array({grapher::json_t{
+              {"type", "regex"},
+              {"pointer", "/name"},
+              {"regex", "*"},
+          }})};
+}
+
+std::vector<predicate_t> get_predicates(group_descriptor_t const &descriptor) {
+  std::vector<predicate_t> predicates;
+
+  predicates.reserve(descriptor.predicates.size());
+  std::ranges::transform(descriptor.predicates, std::back_inserter(predicates),
+                         get_predicate);
+  return predicates;
+}
+
+grapher::json_t::array_t extract_group(group_descriptor_t const &descriptor,
+                                       grapher::json_t::array_t const &events) {
+  std::vector<predicate_t> predicates = get_predicates(descriptor);
+
+  grapher::json_t::array_t res;
+
+  std::ranges::copy_if(
+      events, std::back_inserter(res), [&](grapher::json_t const &event) {
+        return std::ranges::all_of(
+            predicates, [&](predicate_t const &p) { return p(event); });
+      });
+
+  return res;
+}
+
+group_descriptor_t read_descriptor(grapher::json_t const &j) {
+  return {.name = json_at_ref<grapher::json_t::string_t const &>(j, "name"),
+          .predicates =
+              json_at_ref<grapher::json_t::array_t const &>(j, "predicates")};
+}
+
+grapher::json_t group_descriptor_json(group_descriptor_t const &descriptor) {
+  return {
+      {"name", descriptor.name},
+      {"predicates", descriptor.predicates},
+  };
+}
+
+grapher::json_t::array_t
+write_descriptors(std::vector<group_descriptor_t> const &descriptors) {
+  grapher::json_t::array_t res;
+  res.reserve(descriptors.size());
+
+  for (group_descriptor_t const &d : descriptors) {
+    res.push_back(group_descriptor_json(d));
+  }
+  return res;
+}
+
+std::vector<group_descriptor_t>
+read_descriptors(grapher::json_t::array_t const &list) {
+  std::vector<group_descriptor_t> res;
+  res.reserve(list.size());
+  std::transform(list.begin(), list.end(), std::back_inserter(res),
+                 &read_descriptor);
+  return res;
+}
 
 std::vector<double> get_values(benchmark_iteration_t const &iteration,
                                std::vector<predicate_t> const &predicates,
@@ -51,5 +135,69 @@ grapher::json_t merge_into(grapher::json_t a, grapher::json_t const &b) {
   }
   return a;
 }
+
+void save_plot(sciplot::Plot const &plot, std::string const &dest,
+               grapher::json_t const &config) {
+  namespace fs = std::filesystem;
+  std::vector<std::string> plot_file_extensions = config.value(
+      "plot_file_extensions", grapher::json_t::array({".svg", ".png"}));
+
+  // Saving file for all extensions
+  for (std::string const &extension : plot_file_extensions) {
+    fs::path file_dest = dest + extension;
+    fs::create_directories(file_dest.parent_path());
+
+    constexpr std::size_t max_filename_size = 256;
+
+    // Avoid filename hitting OS filename size limit (yes, this is bad)
+    if (file_dest.filename().string().size() > max_filename_size) {
+      std::string new_filename = file_dest.stem();
+      new_filename.resize(max_filename_size - extension.size());
+      file_dest.replace_filename(new_filename);
+    }
+
+    plot.save(dest + extension);
+  }
+}
+
+sciplot::Plot &apply_config(sciplot::Plot &plot, grapher::json_t config) {
+  // Dimensions
+  if (config.contains("width") && config.contains("height")) {
+    plot.size(config["width"], config["height"]);
+  }
+
+  // Labels
+  if (config.contains("legend_title")) {
+    plot.legend().atOutsideRightTop().title(config["legend_title"]);
+  }
+
+  if (config.contains("x_label")) {
+    plot.xlabel(config["x_label"]);
+  }
+
+  if (config.contains("ylabel")) {
+    plot.ylabel(config["y_label"]);
+  }
+
+  return plot;
+}
+
+grapher::json_t const default_config = {
+    {"width", 1500},
+    {"height", 500},
+    {"legend_title", "Timings"},
+    {"x_label", "Benchmark size factor"},
+    {"y_label", "Time (µs)"},
+    {"plot_file_extensions", grapher::json_t::array({".svg", ".png"})},
+};
+
+/// Common plot JSON parameters:
+/// - `width` (`int`): Graph width
+/// - `height` (`int`): Graph height
+/// - `legend_title` (`string`): Graph legend title
+/// - `x_label` (`string`): X axis label
+/// - `y_label` (`string`): Y axis label
+/// - `plot_file_extensions` (string array): List of extensions for the export
+grapher::json_t base_default_config() { return default_config; }
 
 } // namespace grapher
